Extract dedicated lookup and definition processors
Split large processor classes into focused, single-responsibility processors:

- PersistenceUnitLookupProcessor: PU lookup definition (both paradigms),
  with monolithic ComponentLookup decomposed into private helpers
- PersistenceUnitDefinitionBlockingProcessor: blocking PU request
  collection, definition, aggregation, and DS reference production
- PersistenceUnitDefinitionReactiveProcessor: reactive PU request
  collection, definition, DS references, and PU building from config
- DataSourceLookupProcessor: DS lookup from handlers
- DataSourceDefinitionBlockingProcessor: JDBC handler, request
  collection, and definition (from AgroalProcessor)
- DataSourceDefinitionReactiveProcessor: reactive handler, request
  collection, and definition (from ReactiveDataSourceProcessor)

Also moves getMultiTenancyStrategy to HibernateProcessorUtil as a
small public helper shared across processors.
