+++ extensions/datasource/deployment/src/main/java/io/quarkus/datasource/deployment/DataSourceProcessor.java
@@ -5,55 +5,19 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.function.Function;
 
 import io.quarkus.datasource.deployment.spi.DataSourceDbKindResolverBuildItem;
 import io.quarkus.datasource.deployment.spi.DataSourceDefinedBuildItem;
 import io.quarkus.datasource.deployment.spi.DataSourceDefinitionBuildItem;
-import io.quarkus.datasource.deployment.spi.DataSourceLookupBuildItem;
-import io.quarkus.datasource.deployment.spi.DataSourceRequestHandlerBuildItem;
 import io.quarkus.datasource.deployment.spi.DefaultDataSourceDbKindBuildItem;
 import io.quarkus.datasource.runtime.DataSourcesBuildTimeConfig;
 import io.quarkus.deployment.annotations.BuildProducer;
 import io.quarkus.deployment.annotations.BuildStep;
-import io.quarkus.deployment.component.ComponentLookup;
 import io.quarkus.deployment.pkg.builditem.CurateOutcomeBuildItem;
 import io.quarkus.runtime.util.ProgrammingParadigm;
-import io.quarkus.runtime.util.Reason;
 
 class DataSourceProcessor {
     public static final String TEST = "test";
-
-    @BuildStep
-    DataSourceLookupBuildItem defineLookup(List<DataSourceRequestHandlerBuildItem> handlers) {
-        boolean blockingFound = false;
-        boolean reactiveFound = false;
-        Function<String, List<Reason>> blockingUnavailableFunction = ignored -> List
-                .of(new Reason("Agroal extension is absent"));
-        Function<String, List<Reason>> reactiveUnavailableFunction = ignored -> List
-                .of(new Reason("Reactive Datasource extension is absent"));
-        for (DataSourceRequestHandlerBuildItem handler : handlers) {
-            switch (handler.getParadigm()) {
-                case BLOCKING -> {
-                    if (blockingFound) {
-                        throw new IllegalStateException("Multiple blocking datasource request handlers " + handlers);
-                    }
-                    blockingFound = true;
-                    blockingUnavailableFunction = handler.getUnavailableFunction();
-                }
-                case REACTIVE -> {
-                    if (reactiveFound) {
-                        throw new IllegalStateException("Multiple blocking datasource request handlers " + handlers);
-                    }
-                    reactiveFound = true;
-                    reactiveUnavailableFunction = handler.getUnavailableFunction();
-                }
-            }
-        }
-
-        return new DataSourceLookupBuildItem(
-                ComponentLookup.of(blockingUnavailableFunction, reactiveUnavailableFunction));
-    }
 
     @BuildStep
     DataSourceDbKindResolverBuildItem resolveDbKinds(
